refactor: use crate:: relative paths for importing quickselect
Apply changes mentioned in suggestion:
https://github.com/wezterm/wezterm/pull/6677#discussion_r1986084447

This makes it a little easier to refactor in the future
